refactor: rename folder according to convention
