+++ src/app/dorm/[id]/page.tsx
@@ -1,6 +1,6 @@
 import { getDormByID } from '@/action/dorm/getDormByID';
 import { ImageCarousel } from '@/components/dorm-page/imageCarousel';
-import { RequestBtn } from '@/components/lesseeHome-page/requestBtn';
+import { RequestBtn } from '@/components/home-page/requestBtn';
 import React from 'react';
 
 const page = async ({ params }: { params: Promise<{ id: string }> }) => {
