+++ src/components/home-page/propertyCard.tsx
