+++ src/components/home-page/requestBtn.tsx
@@ -8,7 +8,7 @@ import {
   DialogTitle,
   DialogTrigger,
 } from '@/components/ui/dialog';
-import { RequestForm } from '@/components/lesseeHome-page/requestForm';
+import { RequestForm } from '@/components/home-page/requestForm';
 
 export function RequestBtn() {
   return (
